update use case: search matching

+++ trunk/befriends/src/java/business/MatchScoreComparator.java
@@ -13,8 +13,8 @@ import java.util.Comparator;
  */
 public class MatchScoreComparator implements Comparator {
     public int compare(Object obj1, Object obj2) {
-        int score1 = ((AccountOfMatch) obj1).getMatchScore();
-        int score2 = ((AccountOfMatch) obj2).getMatchScore();
+        float score1 = ((AccountOfMatch) obj1).getMatchScore();
+        float score2 = ((AccountOfMatch) obj2).getMatchScore();
         if (score1 > score2) {
             return 1;
         } else if (score1 == score2) {
